Adjust signup notification success styling (#912)
# Adjust signup notification success styling

When users sign up to the site and get redirected, they see the following message

![Screenshot from 2023-07-31 16-47-01.png](/attachments/2f61ab23-f004-4eab-89c3-df133335d262)

This uses a missing `alert__success` class.

I've added this in using the existing `--kbin-success-color`, so this should work on all themes (and themes themselves can re-implement these variables if they prefer)

(success message, as seen on `tokyo night` theme)
![Screenshot from 2023-07-31 17-00-56.png](/attachments/59d64f6e-d16b-4276-9c73-96803a5d65ea)

Anywhere that uses the `_flash.html.twig` template with a success message will now get this styling automatically.

Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/912
Reviewed-by: Melroy van den Berg <melroy89@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>