Add danger button styling to user delete action (#910)
# Add danger button styling to user delete action

[As outlined here](https://codeberg.org/Kbin/kbin-core/issues/870), the **delete account** button blends into the UI a bit and if someone clicks through their account they could accidentially delete their account.

(an easily pressed delete button)
![current UI ](https://codeberg.org/Kbin/kbin-core/attachments/ad3fad01-732e-4255-83fa-4e80b4de7831)

In this PR I've added a new **danger** style for buttons and applied it here. This makes the button much more obvious that it triggers a potentially dangerous action.

![Screenshot from 2023-07-31 10-56-27.png](/attachments/88d02283-7ea8-49da-b4ae-d76f1704d50b)

Eventually to improve this we can move the **delete** action to a dedicated page, but for right now this is an improvement.

Other buttons can now implement the `btn__danger` class also.

Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/910
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>