Adjusted top header links click functionality (#676)
# Adjusted top nav menu click functionality

There are a few [older tickets raised](https://codeberg.org/Kbin/kbin-core/issues/153) that mentioned that the header items (threads, microblog, people) etc were not clickable correctly on their clients.

We have a controller set on the main nav that handles the clicking of the `a` tags and manually adjusts the URL with `window.location`, this is a leftover from when we moved the mobile nav toggle out of that controller.

This functionality isnt needed anymore, these should just be normal links so people can click on them or control click to open them in a new tab.

![the main nav](https://codeberg.org/attachments/edee2a7c-eb17-4450-9040-f990c7e670a4)

This PR removes this old functionality, see the below video below to confirm that clicking through to links still works. Basically we shouldnt interfere with links unless we're doing something special.

(clicks still work)
[Screencast from 10-07-23 14:04:49.webm](/attachments/2863ee5a-9921-4eaf-8259-5c7d0a78d978)

Co-authored-by: AnonymousLlama <simonrcodrington@gmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/676
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>