Adjust mobile design for sidebar top menu (#967)
# Adjust mobile design for sidebar top menu

There's a large gap at the top of the mobile sidebar, where the primary actions for threads, microblog, people etc are displayed

(large top / bottom gap and padding)
![Screenshot from 2023-08-07 17-06-49.png](/attachments/7986e0d2-17d8-4a2b-bba6-568c325d366f)

In this PR I've leveraged our existing styles to bring them more in-line with what our sidebar items look like.

(sidebar when viewing the homepage)
![Screenshot from 2023-08-07 17-05-03.png](/attachments/82605c84-9ec8-42a0-9820-95239107b6f7)

(sidebar when viewing selected pages like a **magazine** or **domain**, we use our similar header layout here)
![Screenshot from 2023-08-07 17-05-11.png](/attachments/b7bfbe38-d399-42b4-99e3-61d2f1e7dfb6)

This change I feel improves the UI on mobile and clearly separates out the **header** on top of the menu in places where it's shown.

Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-authored-by: Szymon Kamiński <kaminskisj@gmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/967
Reviewed-by: Szymon <szsz@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>