Fix broken mobile sidebar issue with the header (#974)
The recent commit here (https://codeberg.org/Kbin/kbin-core/commit/8d7f6881174b2dcf4c7f57145dd3980847ec3258) added some changes to the mobile nav to make it pretty.

The **title** at the top of the menu that sometime shows has a **stretched link** and because the parent isnt **relative** it expands the full width of the page. Meaning when you press forms it clicks the link instead

![Screenshot from 2023-08-08 10-03-47.png](/attachments/f3a43da8-3b56-45b5-b588-b06c14da3d9a)

This quick fix makes this relative and now it works as expected

![Screenshot from 2023-08-08 10-14-14.png](/attachments/42ed5355-31cc-4705-83f4-12ae48405ee5)

Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/974
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>