Fix bug causing doctrine to attempt to get vote IDs from the non-existent vote_id_seq (#992)
Fixes #991

Not necessarily my favorite way to fix this since it creates redundant code that was previously shared by inheriting from the vote class, but I tried only making the Vote class abstract and that didn't fix it. If anyone knows a better way to fix this please let me know.

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/992
Reviewed-by: Szymon <szsz@noreply.codeberg.org>
Co-authored-by: rideranton <rideranton@hotmail.com>
Co-committed-by: rideranton <rideranton@hotmail.com>