Add a 'show password' button for password fields (#942)
# Add a 'show password' button for password fields

As [outlined in this ticket](https://codeberg.org/Kbin/kbin-core/issues/933), a nice UI/UX improvement would be giving users the ability to see their passwords

In this PR I've introduced a **password preview** toggle button.

(shows the new button in action on the login form)
![Screencast from 04-08-23 12_33_55.gif](/attachments/c0df0f21-c669-47c4-ab23-86863d922baf)

This new preview button swaps out the input `type` from `password` to `text` when pressed, giving users an ability to see their passwords.

I've implemented the same functionality across both the **register** and **change password** form (along with a few others were a password is shown)

(updated user register form)
![Screenshot from 2023-08-04 12-36-35.png](/attachments/7ddb76b9-f761-4430-ba64-43e34ce89c47)

(updated user password change form)
![Screenshot from 2023-08-04 12-38-28.png](/attachments/8f71de03-db41-43c5-8c5d-f8b2f7764cdc)

This is done via a new `stimulus component` applied to these fields, along with styling to get these fields showing side-by-side.

Going forward my next step will be to implement a custom `type` so when we need a new `password` in the future in a new form, we can just implement it and all the functionality will be baked in.

closes https://codeberg.org/Kbin/kbin-core/issues/933

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/942
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>