Prevent non logged in users from searching for federated items (users/magazines) (#978)
This will also add configuration option to allow it, if you really want to (default is do not)

Mainly this needs to be looked over and tested (passed testing on dev environment). Also, I'm sure how we are doing translations but, the added config option will need translating, it is only English right now.

Closes: #887

Co-authored-by: r00ty <r00ty@kbin.life>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/978
Co-authored-by: r00ty <r00ty@noreply.codeberg.org>
Co-committed-by: r00ty <r00ty@noreply.codeberg.org>