Added text color adjustments for placeholders (#957)
# Added text color adjustments for placeholders

Currently some of the **ajax** components e.g. the magazine selector are not styled correctly. It's mostly because of the `placeholder` text color hasn't been set.

(currently on prod, the placeholder is a gray color)
![Screenshot from 2023-08-06 14-48-44.png](/attachments/4b1ac21e-4be0-46f4-901b-247f28381d41)

In this PR I've **added in a style for placeholders**, which affects all textarea/input controls. This is done via a new CSS property and adjusted for each of the themes. The idea is the placeholder is darker than the text (so users know when it's been selected)

Here's how the above ajax control looks on the `Tokyo Night` theme, it uses a darker purple when its value has not been selected.

![Screenshot from 2023-08-06 14-48-15.png](/attachments/130255a4-70cf-43b5-ac30-094f4aea265b)

After its selected it uses the normal text color to be brighter)
![Screenshot from 2023-08-06 14-54-10.png](/attachments/0e94de9e-630b-4747-afd1-78643a974e21)

Also look at the add new form below
![Screenshot from 2023-08-06 14-55-56.png](/attachments/3f4f5aab-5467-4f91-ad43-a53f859f6c5a)

This should make the ajax auto-completes and other parts of the site look a little better

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/957
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>