Adjustments to entry title functionality (#1007)
# Adjustments to entry title functionality

There's been a [few discussions about what should happen](https://codeberg.org/Kbin/kbin-core/issues/356) when you click on the entry title.

Here's a summary from how I see it

- **Title should click through to the single article**
- **Title should click through to the target resource** (e.g the url)
- We should add some type of **symbol or extra link** to give users access to the url

In this PR I've adjusted the way these entries are displayed to hopefully find a middle group that most will like.

## Current Approach

Currently when you view an  **article** in lists, you see the following:

![Screenshot from 2023-08-16 09-04-14.png](/attachments/fedec805-cac2-4091-bbf5-4d990854d5d7)

The **title** links to the **single page**. The link that follows it is actually the **domain**. That domain when clicked on takes you to the listing page for all articles matching that domain e.g all the `rollingstone` content

(showing all content from `/d/rollingstone.com`)
![Screenshot from 2023-08-16 09-07-34.png](/attachments/fba117c7-fa39-478b-b9e1-bc95aa739d1c)

Having that domain link is **useful** but it feels like it's a waste of good real estate that could be used to provide access to the **url**

## New Approach

We now show the **url** if applicable (its only seems set if we didn't create local content) in place of the domain.

(shows title + url )
![Screenshot from 2023-08-16 09-12-27.png](/attachments/2c1184c1-cdd0-4632-b36c-8902fb441d22)

In this approach the user can now press the **title** / **image** to get to the single page (unchanged), but now they can also click on the name next to the title to jump directly to the resource.

In the above example we still show the domain `www.wilderman.com` (so the name is kept short), but it links directly to the remote resource e.g. `http://www.wilderman.com/et-et-nam-aut-delectus-non-repellendus-porro.html` (doesn't exist but it would if real)

To retain access to the domain, I've moved this into the `more` dropdown, so users who want to see all articles from that domain still can.

(still retain access to the `domain` listing in the `more` menu)
![Screenshot from 2023-08-16 08-54-57.png](/attachments/cd3eae42-b381-4d46-a993-4abb46cf9c59)

When we get to the **single entry page**, the title itself already used to use the **url**, so that remains unchanged. The domain is still removed and placed in the **more** menu to keep the two views consistent.

(single page view)
![Screenshot from 2023-08-16 09-16-53.png](/attachments/4aefa98d-ccea-42b5-ad24-7f81ceb9fb53)

## Summary

This quick approach should give access to the two things people seemed to want the most, a way to get to the single page and a way to get to the resource itself.

This feels like it resolves the two main issues that have been bothering people, without us having to re-engineer the solution. How do we feel about this potential change?

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1007
Reviewed-by: Adam Cooper <cooperaj@noreply.codeberg.org>
Reviewed-by: Szymon <szsz@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>