Add new Defederated Instances page for transparency (#971)
# New Defederated Instances page for transparency

Currently there's no way for users to know what **instances** are currently banned / defederated with.

To improve transparency I propose in this PR a **new defederation page** that shows all of the instances the target site no longer connects with

(new UI showing a list of defederated instances)
![Screenshot from 2023-08-07 21-33-09.png](/attachments/75a1428c-f618-4b68-9289-f9f07d601b94)

This just shows a simple list of domains we currently dont federate with (or a message if no domains are currently being blocked)

This is accessible **through the footer**, under the privacy policy

(new page link in the footer, in the help section)
![Screenshot from 2023-08-07 21-33-15.png](/attachments/47fe1587-2674-419a-92c0-2f5c960b42ca)

This page is powered by the **federation** tag in the **admin panel**

(a list of all instances we currently ban)
![Screenshot from 2023-08-07 21-33-32.png](/attachments/0e72bda1-1c85-4ac8-acea-9e5e4242c6c8)

To help admins who might not want to show this page. I've created a new setting under **admin panel** > **setting** > **instance**. This will let them toggle this page on or off.

(admins are able to determine if the defederation page will be displayed or not)
![Screenshot from 2023-08-07 21-33-22.png](/attachments/35a23c85-32a6-4dc2-b834-2836da3bdd2a)

This update should provide some more transparency to users as to the content they're connecting with

Fixes https://codeberg.org/Kbin/kbin-core/issues/968

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/971
Reviewed-by: Adam Cooper <cooperaj@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>