Adjustments to federation page to factor in no banned items (#1021)
# Adjustments to federation page to factor in no banned items

I've adjusted the federation page to factor in the situation of what happens when there are no items on the *banned instances* list, currently if you dont have anything in the banned instances list, the `allowedinstances` array isnt populated correctly.

I've also added in a missing closing `</a>` tag for each item

With this change it should show just the allowed instances if no banned instances are added.

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1021
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>