Remove the  'how to use the editor'  button from the editor UI (#855)
# Removing the 'how to use the editor' button

Currently this button below doesn't do anything.

![Screenshot from 2023-07-25 14-49-06.png](/attachments/be85d526-7c47-4313-9db8-ece261908acd)

This button looks like it may have been used a while back as a popup to show information about how to use the editor. Currently in production its just a dead button, users pressing it won't understand why it's not working.

In this PR I've **proposed we remove this button**. If we want to add in a dedicated help button we should look at that, but right now cleaning up the UI to only have functional components is preferred.

**(button removed from the UI)**
![Screenshot from 2023-07-25 14-49-19.png](/attachments/2be7695c-2aa6-4e60-bb1a-d9b36c65c10d)

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/855
Reviewed-by: Szymon <szsz@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>