Update to sha hash of symfony/messenger until tagged release available (#1027)
FIxes #1026

The php amqp module releases it's first new version in 2 years 2 days ago. It's got a breaking change in it for the symonfy/amqp-messenger component. This PR updates to the untagged commit where this was fixed.

An alternative would be to pin the amqp module at version 1.11.0 in the docker files and update the baremetal install instructions to cover the change.

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1027
Reviewed-by: Szymon <szsz@noreply.codeberg.org>
Co-authored-by: Adam Cooper <adam@acpr.dev>
Co-committed-by: Adam Cooper <adam@acpr.dev>