Adjust Link functionality to improve title / description display (#1041)
# Adjustments to Link entry type

This PR introduces a few UX improvements to the **LINK** entry type.

Currently when you paste into the **url** field, the system uses an ajax request to try and fetch the **title** and **description** of the content to auto-populate the associated fields.

While this works, visually when you paste into the URL field, there's no visual indication of what's happening. When you're on **kbin.social** or other smaller websites, the ajax call can sometimes take 3+ seconds to resolve.

## Style adjustments

The current controller **disables** and **enables** the title / description field as we make these requests, but we dont have any **styling** for these fields so the user can't tell the difference. I've added in **simple disabled styling as seen below**

(visually distinct styling for disabled elements)
![Screenshot from 2023-08-25 10-32-21.png](/attachments/d44e8ee2-12fb-44ce-bc71-153177deed46)

## Swap from `debounce` to `throttle`

Another issue we have is that we're using `debounce` on this control, that means there's a minimum `800ms` delay from when the user pastes into the URL field (contributing to the slowness) and when the ajax call triggers. I've swapped this for `1000ms` of `throttle`, meaning as soon as we've pasted into this field it triggers (but no more than 1 request per second)

## New loading indicator

 I've adjusted the label next to the **url** field and added in the `loading indicator`. This is recycled from another part of the site (and adjusted in size). We'll show this when we're waiting for the ajax call to return.

Visually now there's a strong indicator as to what's going on behind the scenes

(showing the loading indicator + visually re-styling the disabled fields makes the UX more consistent)
![Screencast from 25-08-23 10_23_31.gif](/attachments/eff04a25-13b5-4a2f-9186-391252716ab8)

Fixes https://codeberg.org/Kbin/kbin-core/issues/1035

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1041
Reviewed-by: Melroy van den Berg <melroy89@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>