Added ability to re-send the users email address verification (#936)
# Re-send email address verification

As [mentioned in this ticket](https://codeberg.org/Kbin/kbin-core/issues/929), currently once you've been sent the account activation email, if you don't activate your account quickly, you can get your account stuck in a locked state without any way to request another activation email.

In this PR I've added in a new **re-send account activation email** system.

Displayed at the bottom of the **login** page (and a few others), we now have a new option.

(new option shown at the bottom of the page)
![Screenshot from 2023-08-03 16-18-03.png](/attachments/59284c8f-0d20-44a3-bddf-3fa2f0bead4f)

When a user selects this, they'll go to another page where they can enter the **email** address of their account.

(users can now enter an email and re-send their activation email)
![Screenshot from 2023-08-03 16-19-29.png](/attachments/99d6668a-7d3b-4eed-98fb-32a3f86cb662)

When submitted, behind the scenes we query for the user associated with that email. We perform some basic checks like making sure the **user exists** and that the **account is not already verified**. If they fail (or if we have issues sending the email) we show the following

(error shown during the process)
![Screenshot from 2023-08-03 16-19-05.png](/attachments/9c799fa6-e449-429f-acd7-f68f8d955a3d)

If the user has been found and they aren't verified, we hook into the **same existing system the user signup system uses.** We leverage this and send out the same email again.

(On success, we let the user know that they've been sent a new email)
![Screenshot from 2023-08-03 18-51-46.png](/attachments/12d11660-8e8a-4d0e-a237-20ddeae45c10)

Once they've verified their account they'll be able to log in and all is well.

Closes #281 and closes #929

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/936
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>