Add custom CSS support to user settings (#1024)
# Add custom CSS support to user settings

This PR adds the ability to define custom CSS on the **user** level, in the same way that we handle custom CSS on the **magazine** level.

![Screenshot from 2023-08-21 10-05-32.png](/attachments/b943747e-4114-471c-8510-f7e3fe3b5f37)

The name / help text will need to be updated in `weblate`.

![Screenshot from 2023-08-21 10-03-48.png](/attachments/1c5c817c-1a9d-4e7c-88a4-79330af27232)

As seen above, when we've added in custom styles they appear on the front-end output

Resolves https://codeberg.org/Kbin/kbin-core/issues/316

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1024
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>