Add 'Edited At' in entry meta to distinguish edited entries (#778)
# Add 'edited at' support to entries to distinguish when an entry has been updated

[As outlined in this issue](https://codeberg.org/Kbin/kbin-core/issues/679), it would be handy if we could show an indicator that an **entry** has been edited.

In this PR I've updated the date template to display if the **entry** was edited and how long ago it was.

(we show the edited date next to the published date if applicable)

![entries now have edited date displayed](https://codeberg.org/attachments/049f6443-cf6a-4db4-a512-0d4a6bc4c0e3)

We already have access to the *editedAt* data on the entry, so we reference the existence of this and if it exist we show this output.

The design is a similar approach to how reddit does it as seen below

![Reddit approach](https://codeberg.org/attachments/457aa0f9-3cf1-4a6e-a9d9-9ec8d6105573)

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/778
Co-authored-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>
Co-committed-by: AnonymousLlama <anonymousllama@noreply.codeberg.org>