Add ability for instance admins to view when users were last active via the admin panel (#1086)
Instead of drudging thru the raw database in order to generate reports of when users were last active, I've added the necessary column to the users page in the admin panel. We may need to revisit the backend on how this metric is being generated, but to the first order this should give an admin a rough idea of which registered users are actively using an instance.

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1086
Reviewed-by: Melroy van den Berg <melroy89@noreply.codeberg.org>
Co-authored-by: debounced <debounced@noreply.codeberg.org>
Co-committed-by: debounced <debounced@noreply.codeberg.org>