Add REST API pt. 8 - Add Search API (#1091)
### This PR is one of a series of PRs adding the API in chunks to help aid the review process!
The previous PR is [here](https://codeberg.org/Kbin/kbin-core/pulls/1067)

Changes in this PR:
 * Add Search API that can find users, magazines, and content in the same way as the existing search page
 * Add functional tests exercising the search API in multiple different ways

Note:
The tests for finding remote users and magazines rely on kbin.social being up. This should probably be changed in the future to rely either on mocking the request and response, or by spinning up a second instance on the machine running the tests.

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1091
Co-authored-by: rideranton <rideranton@hotmail.com>
Co-committed-by: rideranton <rideranton@hotmail.com>