Reposition add new comment form on mobile (#525)
# Repositioning the comment submit form for mobile

Currently on mobile the **add comment** form is at the bottom of the page, under the current list of **comments**. Scrolling this list on mobile can be pretty annoying.

**(Having to scroll through multiple comments to get to the end)**
![Screenshot from 2023-06-29 15-36-03.png](/attachments/7ce9885e-8ef5-4f8d-ae2d-60759bfd817c)

What I propose in this PR is that on **mobile** only we move the comment box to the top of the page, directly above the **comment list** as seen below

**(Now appears above the comment list)**
![Screenshot from 2023-06-29 15-15-29.png](/attachments/2267de79-87ee-45ae-8324-508188f1dd00)

This is done using CSS position orders so none of the posting functionality should be affected.

There was an ongoing discussion about why it was originally placed at the bottom (to encourage you to read the content first), but I'd argue that on mobile there's too much scrolling and it detracts from the experience.

Co-authored-by: Simon Codrington <simonrcodrington@gmail.com>
Co-authored-by: AnonymousLlama <simonrcodrington@gmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/525
Reviewed-by: Melroy van den Berg <melroy89@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>