Disable avatars on comments via new user setting (#508)
# Disabling avatar output on the comment level

Currently when we're on the single article page, each comment outputs the users avatar image (if uploaded) as seen below

(several comments on an article)
![Screenshot from 2023-06-27 14-25-59.png](/attachments/afce6cee-c1e3-453e-a8a2-18f5058b9032)

While that looks nice sometimes, often people will come through as empty squares (either they've not uploaded a user name or the federation process has made the images not come through)

## Proposed new setting / section for comments

In this PR I've added a new section in the **settings** sidebar, one for **comments**

(new setting section)
![Screenshot from 2023-06-27 14-19-26.png](/attachments/62b0aa5c-0d62-4709-99ee-731d8b14073a)

The new setting added here is for **Show user's avatar in comments**, which is toggled **on** by default (to ensure images are displayed when not selected, as I assume most would like images)

## Removal of images from comments

When we disable avatar images here, the image output is removed. I feel this is a more clean layout option since you don't have to worry about missing images.

(with avatars disabled on the comment level)
![Screenshot from 2023-06-27 14-30-14.png](/attachments/6ef55be1-17f5-41f5-8cb9-df8312b30a53)

## Feedback and thoughts

I understand we dont want to have too many options here, but I feel this one brings some value, letting users configure how they see comments.

This new settings section also we could use later on if we want to add various layouts for the comments.

As part of the PR i've used Google translate for the translations for the setting, feel free to comment if these need to be adjusted.

Co-authored-by: Simon Codrington <simonrcodrington@gmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/508
Reviewed-by: Melroy van den Berg <melroy89@noreply.codeberg.org>
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>