Adding user about field to profile display (#640)
# Adding user about field to profile display

As outlined in this [ticket here](https://codeberg.org/Kbin/kbin-core/issues/609), it would be nice if we could display the users **about** information.

There's actually several places in the website where the **user profile** card is displayed. This PR seeks to add the about info to each of them where appropriate.

(in the sidebar, when viewing a post from the user for example)
![Screenshot from 2023-07-07 16-40-27.png](/attachments/d49aecfb-77da-439f-bead-0aedb3d7f510)

(when viewing another users profile)
![Screenshot from 2023-07-07 16-40-14.png](/attachments/fdefdc61-d45c-4482-8c4c-265e889650a7)

(when viewing a user in a list of other users)
![Screenshot from 2023-07-07 16-40-43.png](/attachments/3f68e673-2bd9-4d6e-9983-1406eb6e3377)

(when a user is viewing their own profile)
![Screenshot from 2023-07-07 18-37-52.png](/attachments/8147e693-76cb-489e-b3bc-f8b4fbfd7f20)

(when viewing the user popover)
![Screenshot from 2023-07-07 18-38-50.png](/attachments/ab6fd0b3-f21b-46ce-817c-60ed99a7248a)

## Changes to the popover

on mobile I've also re-styled the popover, on mobile this popover now fits the width of the screen, making it at least readable. Functionality wise though it's still **hard to trigger** and **vanishes** when you scroll sometimes. A topic on changing this thing is a matter for another time I feel.

| Before                                                                                                                                                         | After                                                                                                                                                                             |
|----------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
| (on prod the container sticks out the side and doesnt really fit)![Screenshot from 2023-07-07 18-45-14.png](/attachments/0ebc29a0-4d70-4025-a381-12fd4aacc838) | (on mobile thing now fits the width of the screen and doesnt overflow like on prod)![Screenshot from 2023-07-07 18-44-25.png](/attachments/a6475057-03cb-40ac-91c9-4c86938b613e)
 |

## Additional thoughts

Currently we output the full text of the **abut** field in these templates. The user can have up to 512 chars in that field.

I'd like feedback on **how** we display this data. Right now we can output the raw HTML and that looks good on the **profile** page for the user. But when we view the **popover** or **sidebar**, do we want to **trim** the content so it better fits.

If we want one area to show the **full** content, and another to be a cut down version, I'll need to make a filter so we can strip out the HTML and trim it to 150 chars etc. However, some people might be using headers, bullet points or other marke-down that might make reading their about us strange.

Co-authored-by: AnonymousLlama <simonrcodrington@gmail.com>
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/640
Co-authored-by: AnonymousLlama <someguy12345678@protonmail.com>
Co-committed-by: AnonymousLlama <someguy12345678@protonmail.com>