Fix #702: Better handling of Sensitive Media. [Untested] (#905)
See my analysis in #702. I haven't tested this code, I only refactored the checking of `$object['sensitive']` into a helper function and applied it to all of the `create` helper functions in `Note.php`. I also applied `toLowerCase` to `adultBadge.textContent` in `subject_controller.js`, which I consider to be Defensive Programming but wouldn't deny anyone calling it superstitious.

I also want to note that this fix may only apply to the ontologies currently defined or recognized by Kbin: Mastodon, Pleroma, Lemmy, and Peertube. This fix will probably NOT work for content federated from Misskey or Calckey instances - those platforms use a property called `isSensitive` and apply the property at the _image_ level, not the _post_ level. Federation with those platforms is currently tenuous at best anyway, but in the future it should be tackled.

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/905
Reviewed-by: Melroy van den Berg <melroy89@noreply.codeberg.org>
Co-authored-by: mpm11011 <mark.mauriello@uconn.edu>
Co-committed-by: mpm11011 <mark.mauriello@uconn.edu>