/kbin RTR#5 Transfer request for ownership of abandoned magazines (#1216)
This update introduces the ability to request the takeover of abandoned magazines. The karab.in instance has been updated to the latest version of the code, which prepares for the kbin.social update after some tests.

- list of abandoned magazines (owner hasn't logged in for a month)
- the ability to submit a request for taking ownership of a warehouse @todo
s- ection in the administrator panel where one can reject or confirm the takeover of a magazine @todo

Time remaining for the kbin.social update: ? days.

[/kbin Road to Release](https://kbin.social/m/kbinMeta/t/588413/kbin-Road-to-Release-devlog-series)

Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1216
Co-authored-by: Ernest Wiśniewski <ernestwisniewski2@gmail.com>
Co-committed-by: Ernest Wiśniewski <ernestwisniewski2@gmail.com>